Change code location to application root
